fixed user management and added default scoring / ranges button to go back to the default vaules

+++ src/main/java/edu/hm/wedoit/rest/UserController.java
@@ -8,10 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -116,8 +113,13 @@ public class UserController
         }
     }
 
-    public ResponseEntity<String> deleteUser(@RequestParam String username){
+    @RequestMapping(value = "/delete/{username:.+}", method = RequestMethod.DELETE)
+    public ResponseEntity<String> deleteUser(@PathVariable String username){
         logger.debug("deleteUser {} " + username);
+
+        if(username.equals("admin")) {
+            return new ResponseEntity<String>("Could not delete the admin", HttpStatus.CONFLICT);
+        }
         if(userService.removeUser(username))
         {
             logger.debug("deleted {}",username);
@@ -127,8 +129,21 @@ public class UserController
         else
         {
             logger.debug("not deleted {}",username);
-            ResponseEntity<String> responseEntity = new ResponseEntity<String>(HttpStatus.CONFLICT);
+            ResponseEntity<String> responseEntity = new ResponseEntity<String>("Cannot delete user with name " + username,
+                    HttpStatus.CONFLICT);
             return responseEntity;
+        }
+    }
+
+    @RequestMapping(value = "/edit" , method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
+    public ResponseEntity<String> editUser(@RequestBody User user) {
+        logger.debug("edit user " + user.getName());
+        if(userService.updateUser(user)) {
+           logger.debug("Edit");
+           return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else {
+           return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
     }
